+++ code/core/src/main/java/io/zabrek/soulbound/api/kernel/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Essential core functionality of SoulBound.
+ */
+package io.zabrek.soulbound.api.kernel;