+++ code/core/src/main/java/io/zabrek/soulbound/kernel/components/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The building blocks of SoulBound.
+ */
+package io.zabrek.soulbound.kernel.components;