+++ code/core/src/main/java/io/zabrek/soulbound/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * Root package for the SoulBound plugin.
+ */
+@DefaultQualifier(value = NotNull.class, locations = {FIELD, PARAMETER, RETURN})
+package io.zabrek.soulbound;
+
+import org.checkerframework.framework.qual.DefaultQualifier;
+import org.jetbrains.annotations.NotNull;
+
+import static org.checkerframework.framework.qual.TypeUseLocation.FIELD;
+import static org.checkerframework.framework.qual.TypeUseLocation.PARAMETER;
+import static org.checkerframework.framework.qual.TypeUseLocation.RETURN;